Introduce Iooxa.dev
I have just released https://iooxa.dev which is a suite of components and layouts to make creating interactive and reactive writing on the web easier.

There are tangle-like components, common-inputs, basic reactive-SVG components, as well as layouts, equations, callouts etc.  for common articles templates.

Let me know if you want any other details, or check out the docs at https://iooxa.dev, hopefully it is a good fit for this page! :)

Not sure if the ordering is important - I put it close to Idyll, as it is pretty similar. See https://github.com/iooxa/article/issues/8 for a discussion!